iss: Refactor translation generator dispatching
- Introduces `DirectGvecTranslateGenerator` and `ScalarTcgTranslateGenerator` for improved modularity and explicit strategy handling.
- Implements a unified `InstructionTranslateGenerator` interface.
- Updates `IssTranslateCodeGenerator` to dispatch generators based on execution plans.
- Adds test coverage to validate correct generator selection for scalar, vector, and helper strategies.
